move all the custom prop validators to single file
