+++ src/components/icon/SecondaryIcon.jsx
@@ -19,50 +19,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 
-import { COLORS, SECONDARY_ICONS, SIZES } from '../../utils/constants';
+import { nameOrText } from '../../utils/propValidators';
+import { COLORS, SIZES } from '../../utils/constants';
 
 import './icons/icons.svg';
 import './icon.less';
-
-/* eslint-disable no-unused-vars */
-// Our custom props validator. I am not in love with the name, but a descriptive name was absurdly long.
-function nameOrText(props, propName, componentName = 'SecondaryIcon') {
-  const types = ['name', 'text'];
-  const otherProp = types.filter(e => e !== propName)[0]; // filter returns an array, I just need the element.
-
-  let error;
-
-  // First make sure we have at least one of them.
-  if (!props[propName] && !props[otherProp]) {
-    error = new Error(
-      `${propName} and ${otherProp} are blank. You must provide one or the other.`,
-    );
-  }
-
-  // If we have a name, we need to make sure it is in the list.
-  if (props.name && !SECONDARY_ICONS.includes(props.name)) {
-    error = new Error(
-      `Invalid prop: name. Prop must be one of ${SECONDARY_ICONS.join(', ')}.`,
-    );
-  }
-
-  // If we have text, it can't be longer than 2 characters.
-  if (props.text && props.text.length > 2) {
-    error = new Error(
-      `Invalid prop: text. Prop cannot be longer than 2 characters.`,
-    );
-  }
-
-  // And lastly, you can have one OR the other, not both.
-  if (props[propName] && props[otherProp]) {
-    error = new Error(
-      `${propName} and ${otherProp} both have values. You must provide one or the other, but not both`,
-    );
-  }
-
-  return error;
-}
-/* eslint-enable no-unused-vars */
 
 /**
  * These light icons are mostly used as an actionable element.
